Our structured cloning implementation does not encode all of RegExp's flags https://bugs.webkit.org/show_bug.cgi?id=227779
Reviewed by Yusuke Suzuki.

LayoutTests/imported/w3c:

Rebaseline WPT test.

* web-platform-tests/IndexedDB/structured-clone.any-expected.txt:

Source/WebCore:

Our structured cloning implementation does not encode all of RegExp's flags.
In particular, we are failing to encode the 'y', 'u', 's' and 'd' flags.

This is causing one of the checks to fail in WebKit on:
- http://wpt.live/IndexedDB/structured-clone.any.html?61-80

That subtest is passing in Firefox and Chrome.

Test: fast/dom/Window/post-message-regexp.html

* bindings/js/SerializedScriptValue.cpp:
(WebCore::CloneSerializer::dumpIfTerminal):

LayoutTests:

Add layout test coverage.

* fast/dom/Window/post-message-regexp-expected.txt: Added.
* fast/dom/Window/post-message-regexp.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@279706 268f45cc-cd09-0410-ab3c-d52691b4dbfc