[GTK] Support dark scrollbars https://bugs.webkit.org/show_bug.cgi?id=227834
Patch by Alexander Mikhaylenko <alexm@gnome.org> on 2021-07-09
Reviewed by Michael Catanzaro.

Also adjust the colors in light variant to be closer to the GTK scrollbars, and
remove an unused variable.

Unfortunately this leaves the scrollbar corner when non-overlay scrollbars are used,
but it was already broken for light scrollbars as well.

* platform/adwaita/ScrollbarThemeAdwaita.cpp:
(WebCore::ScrollbarThemeAdwaita::updateScrollbarOverlayStyle):
(WebCore::ScrollbarThemeAdwaita::paint):
* platform/adwaita/ScrollbarThemeAdwaita.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@279782 268f45cc-cd09-0410-ab3c-d52691b4dbfc