[CG] REGRESSION(r278863): The destination rectangle is truncated when the sub-image is used https://bugs.webkit.org/show_bug.cgi?id=227614 <rdar://79840643>
Reviewed by Simon Fraser.

Source/WebCore:

This patch gets the calculation of the destRect in the case of the sub-
image as it was before r278863.

The size of the destRect has to be equal to the backend size of the
ImageBuffer in logical coordinates.

* platform/graphics/cg/GraphicsContextCG.cpp:
(WebCore::GraphicsContextCG::drawNativeImage):

Tools:

Add an API test to test drawing an ImageBuffer into another and both
have the logicalSize scaled such that they have pixels.

* TestWebKitAPI/Tests/WebCore/ImageBufferTests.cpp:
(TestWebKitAPI::TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@279885 268f45cc-cd09-0410-ab3c-d52691b4dbfc